+++ dbos_transact/system_database.py
@@ -3,7 +3,7 @@ import select
 import threading
 import time
 from enum import Enum
-from typing import Any, Dict, Literal, Optional, Sequence, TypedDict, cast
+from typing import Any, Dict, List, Literal, Optional, Sequence, TypedDict, cast
 
 import psycopg2
 import sqlalchemy as sa
@@ -16,11 +16,6 @@ from dbos_transact.error import (
     DBOSDuplicateWorkflowEventError,
     DBOSNonExistentWorkflowError,
     DBOSWorkflowConflictUUIDError,
-)
-from dbos_transact.workflow import (
-    GetWorkflowsInput,
-    GetWorkflowsOutput,
-    WorkflowInformation,
 )
 
 from .dbos_config import ConfigFile
@@ -76,47 +71,47 @@ class GetEventWorkflowContext(TypedDict):
     timeout_function_id: int
 
 
+class GetWorkflowsInput:
+    name: Optional[str]  # The name of the workflow function
+    authenticated_user: Optional[str]  # The user who ran the workflow.
+    start_time: Optional[str]  # Timestamp in ISO 8601 format
+    end_time: Optional[str]  # Timestamp in ISO 8601 format
+    status: Optional[WorkflowStatuses]
+    application_version: Optional[
+        str
+    ]  # The application version that ran this workflow.
+    limit: Optional[
+        int
+    ]  # Return up to this many workflows IDs. IDs are ordered by workflow creation time.
+
+
+class GetWorkflowsOutput:
+    def __init__(self, workflow_uuids: List[str]):
+        self.worfkflow_uuids = workflow_uuids
+
+    workflow_uuids: List[str]
+
+
+class WorkflowInformation(TypedDict, total=False):
+    workflow_uuid: str
+    status: WorkflowStatuses  # The status of the workflow.
+    name: str  # The name of the workflow function.
+    workflow_class_name: str  # The class name holding the workflow function.
+    workflow_config_name: (
+        str  # The name of the configuration, if the class needs configuration
+    )
+    authenticated_user: str  # The user who ran the workflow. Empty string if not set.
+    assumed_role: str
+    # The role used to run this workflow.  Empty string if authorization is not required.
+    authenticatedRoles: List[str]
+    # All roles the authenticated user has, if any.
+    input: Optional[WorkflowInputs]
+    output: Optional[str]
+    error: Optional[str]
+    request: Optional[str]
+
+
 dbos_null_topic = "__null__topic__"
-
-"""
-
-export async function listWorkflows(config: DBOSConfig, input: GetWorkflowsInput, getRequest: boolean) {
-  const systemDatabase = new PostgresSystemDatabase(config.poolConfig, config.system_database, createLogger() as unknown as GlobalLogger)
-  const workflowUUIDs = (await systemDatabase.getWorkflows(input)).workflowUUIDs.reverse(); // Reverse so most recent entries are printed last
-  const workflowInfos = await Promise.all(workflowUUIDs.map(async (i) => await getWorkflowInfo(systemDatabase, i, getRequest)))
-  await systemDatabase.destroy();
-  return workflowInfos;
-}
-
-export async function getWorkflow(config: DBOSConfig, workflowUUID: string, getRequest: boolean) {
-  const systemDatabase = new PostgresSystemDatabase(config.poolConfig, config.system_database, createLogger() as unknown as GlobalLogger)
-  const info = await getWorkflowInfo(systemDatabase, workflowUUID, getRequest);
-  await systemDatabase.destroy();
-  return info;
-}
-
-// Cancelling a workflow prevents it from being automatically recovered, but active executions are not halted.
-export async function cancelWorkflow(config: DBOSConfig, workflowUUID: string) {
-  const systemDatabase = new PostgresSystemDatabase(config.poolConfig, config.system_database, createLogger() as unknown as GlobalLogger)
-  await systemDatabase.setWorkflowStatus(workflowUUID, StatusString.CANCELLED, false)
-  await systemDatabase.destroy();
-}
-
-export async function reattemptWorkflow(config: DBOSConfig, runtimeConfig: DBOSRuntimeConfig | null, workflowUUID: string, startNewWorkflow: boolean) {
-  const dbosExec = new DBOSExecutor(config);
-  if (runtimeConfig !== null) {
-    await DBOSRuntime.loadClasses(runtimeConfig.entrypoints);
-  }
-  await dbosExec.init();
-  if (!startNewWorkflow) {
-    await dbosExec.systemDatabase.setWorkflowStatus(workflowUUID, StatusString.PENDING, true);
-  }
-  const handle = await dbosExec.executeWorkflowUUID(workflowUUID, startNewWorkflow);
-  const output = await handle.getResult();
-  await dbosExec.destroy();
-  return output;
-}
-"""
 
 
 class SystemDatabase:
